+++ bw/math/Matrix.tpp
@@ -9,7 +9,7 @@ namespace bw
 {
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T>::Matrix()
         : _data(std::make_unique<T[]>(R * C))
     {
@@ -18,13 +18,13 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T>::Matrix(T** initializer)
         : _data(std::make_unique<T[]>(R* C))
     {
-        for (size_t i = 0; i < R; i++)
+        for (int i = 0; i < R; i++)
         {
-            for (size_t j = 0; j < C; j++)
+            for (int j = 0; j < C; j++)
             {
                 (*this)(i, j) = initializer[i][j];
             }
@@ -33,13 +33,13 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T>::Matrix(const std::array<std::array<T, C>, R>& data)
         : _data(std::make_unique<T[]>(R* C))
     {
-        for (size_t i = 0; i < R; ++i)
+        for (int i = 0; i < R; ++i)
         {
-            for (size_t j = 0; j < C; ++j)
+            for (int j = 0; j < C; ++j)
             {
                 (*this)(i, j) = data[i][j];
             }
@@ -48,19 +48,19 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T>::Matrix(const std::vector<std::vector<T>>& initializer)
         : _data(std::make_unique<T[]>(R* C))
     {
         if (initializer.size() != R)
             throw std::invalid_argument("Invalid number of rows");
 
-        for (size_t i = 0; i < R; i++)
+        for (int i = 0; i < R; i++)
         {
             if (initializer[i].size() != C)
                 throw std::invalid_argument("Invalid number of columns in row " + std::to_string(i));
 
-            for (size_t j = 0; j < C; j++)
+            for (int j = 0; j < C; j++)
             {
                 (*this)(i, j) = initializer[i][j];
             }
@@ -69,7 +69,7 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T>::Matrix(const Matrix<R, C, T>& other)
         : _data(std::make_unique<T[]>(R* C))
     {
@@ -78,8 +78,8 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
-    T& Matrix<R, C, T>::operator()(size_t row, size_t col)
+    template<int R, int C, typename T>
+    T& Matrix<R, C, T>::operator()(int row, int col)
     {
         if (row >= R || col >= C)
             throw std::out_of_range("Matrix indices out of range: [" +
@@ -90,8 +90,8 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
-    const T& Matrix<R, C, T>::operator()(size_t row, size_t col) const
+    template<int R, int C, typename T>
+    const T& Matrix<R, C, T>::operator()(int row, int col) const
     {
         if (row >= R || col >= C)
             throw std::out_of_range("Matrix indices out of range: [" +
@@ -102,8 +102,8 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
-    T* Matrix<R, C, T>::operator[](size_t row)
+    template<int R, int C, typename T>
+    T* Matrix<R, C, T>::operator[](int row)
     {
         if (row >= R)
             throw std::out_of_range("Matrix row index out of range: " +
@@ -113,8 +113,8 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
-    const T* Matrix<R, C, T>::operator[](size_t row) const
+    template<int R, int C, typename T>
+    const T* Matrix<R, C, T>::operator[](int row) const
     {
         if (row >= R)
             throw std::out_of_range("Matrix row index out of range: " +
@@ -124,15 +124,15 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     template<typename U>
     Matrix<R, C, T>::operator Matrix<R, C, U>() const
     {
         Matrix<R, C, U> converted;
 
-        for (size_t i = 0; i < R; ++i)
+        for (int i = 0; i < R; ++i)
         {
-            for (size_t j = 0; j < C; ++j)
+            for (int j = 0; j < C; ++j)
             {
                 converted(i, j) = static_cast<U>((*this)(i, j));
             }
@@ -143,10 +143,10 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     bool Matrix<R, C, T>::operator==(const Matrix<R, C, T>& other) const
     {
-        for (size_t i = 0; i < R * C; ++i)
+        for (int i = 0; i < R * C; ++i)
         {
             if (_data[i] != other._data[i])
                 return false;
@@ -156,7 +156,7 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     bool Matrix<R, C, T>::operator!=(const Matrix<R, C, T>& other) const
     {
         return !(*this == other);
@@ -164,7 +164,7 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T>& Matrix<R, C, T>::operator=(const Matrix<R, C, T>& other)
     {
         if (this != &other)
@@ -176,12 +176,12 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T> Matrix<R, C, T>::operator+(const Matrix& other) const
     {
         Matrix<R, C, T> result;
 
-        for (size_t i = 0; i < R * C; ++i)
+        for (int i = 0; i < R * C; ++i)
         {
             result._data[i] = _data[i] + other._data[i];
         }
@@ -191,12 +191,12 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T> Matrix<R, C, T>::operator-(const Matrix& other) const
     {
         Matrix<R, C, T> result;
 
-        for (size_t i = 0; i < R * C; ++i)
+        for (int i = 0; i < R * C; ++i)
         {
             result._data[i] = _data[i] - other._data[i];
         }
@@ -206,12 +206,12 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     Matrix<R, C, T> Matrix<R, C, T>::operator*(T scalar) const
     {
         Matrix<R, C, T> result;
 
-        for (size_t i = 0; i < R * C; ++i)
+        for (int i = 0; i < R * C; ++i)
         {
             result._data[i] = _data[i] * scalar;
         }
@@ -221,16 +221,16 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
-    template<size_t Arg_C>
+    template<int R, int C, typename T>
+    template<int Arg_C>
     Matrix<R, Arg_C, T> Matrix<R, C, T>::operator*(const Matrix<C, Arg_C, T>& other) const
     {
         Matrix<R, Arg_C, T> result;
 
-        for (size_t i = 0; i < R; i++) {
-            for (size_t k = 0; k < Arg_C; k++) {
+        for (int i = 0; i < R; i++) {
+            for (int k = 0; k < Arg_C; k++) {
                 T sum{};
-                for (size_t j = 0; j < C; j++) {
+                for (int j = 0; j < C; j++) {
                     sum += (*this)(i, j) * other(j, k);
                 }
                 result(i, k) = sum;
@@ -242,10 +242,10 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     void Matrix<R, C, T>::fill(T value)
     {
-        for (size_t i = 0; i < R * C; ++i)
+        for (int i = 0; i < R * C; ++i)
         {
             _data[i] = value;
         }
@@ -253,12 +253,12 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     void Matrix<R, C, T>::translate(Vec3<T> position) requires std::is_floating_point_v<T> && (R == 4) && (C == 4)
     {
         Matrix<4, 4, T> translation;
 
-        for (size_t i = 0; i < 4; ++i)
+        for (int i = 0; i < 4; ++i)
             translation(i, i) = static_cast<T>(1);
 
         translation(0, 3) = position.x;
@@ -270,7 +270,7 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     std::array<T, C * R> Matrix<R, C, T>::toLinearArray() const
     {
         std::array<T, R* C> arr;
@@ -280,15 +280,15 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
-    size_t Matrix<R, C, T>::_getIndex(size_t row, size_t col) const
+    template<int R, int C, typename T>
+    int Matrix<R, C, T>::_getIndex(int row, int col) const
     {
         return row * C + col;
     }
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     void Matrix<R, C, T>::_copyFrom(const Matrix& other)
     {
         if constexpr (std::is_trivially_copyable_v<T>)
@@ -297,7 +297,7 @@ namespace bw
         }
         else
         {
-            for (size_t i = 0; i < R * C; ++i)
+            for (int i = 0; i < R * C; ++i)
             {
                 _data[i] = other._data[i];
             }
@@ -306,13 +306,13 @@ namespace bw
 
     ////////////////////////////////////////////////////////////
 
-    template<size_t R, size_t C, typename T>
+    template<int R, int C, typename T>
     std::ostream& operator<<(std::ostream& os, const Matrix<R, C, T>& matrix)
     {
-        for (size_t i = 0; i < R; ++i)
+        for (int i = 0; i < R; ++i)
         {
             os << "[ ";
-            for (size_t j = 0; j < C; ++j)
+            for (int j = 0; j < C; ++j)
             {
                 os << matrix(i, j);
                 if (j + 1 < C) os << ", ";
